Refactor

+++ instance/startup.js
@@ -18,8 +18,6 @@ const build = image => {
 };
 
 let child;
-let appRef;
-let pid;
 
 rp({ json: true, uri: metaDataUrl }).then(content => {
   const instanceId = content.instanceId;
@@ -33,43 +31,35 @@ rp({ json: true, uri: metaDataUrl }).then(content => {
   const appId = proc.execSync(getApp, { encoding: 'utf-8' }).replace('\n', '');
   const userId = proc.execSync(getUser, { encoding: 'utf-8' }).replace('\n', '');
 
-  // create the Firebase ref
-  appRef = new Firebase(`https://restle-launch2016.firebaseio.com/apps/${appId}`);
-
   process.env.USER_ID = userId;
   process.env.APP_ID = appId;
   process.env.PORT = 8000;
 }).then(() => {
-  let appName = '';
+  const appRef = new Firebase(`https://restle-launch2016.firebaseio.com/apps/${process.env.APP_ID}`);
 
-  appRef.on('value', snapshot => {
-    appName = snapshot.val().name;
-    console.log('appName:', appName);
-  });
+  // if image changes value
+  // if image exists, reinstantiate restle
 
   // when the app image changes, reinstantiate the restle app
   appRef.child('image').on('value', snapshot => {
     appRef.child('isDeploying').set(true);
-    console.log('isDeploying:', true);
 
-    build(snapshot.val());
+    if (snapshot.exists()) {
+      // build `dist`
+      build(snapshot.val());
 
-    // if the instance is already running, restart it
-    if (child !== undefined) {
-      console.log('Time to disconnect child!');
-      child.kill('SIGKILL');
+      // if the instance is already running, restart it
+      if (child !== undefined) {
+        console.log('Time to disconnect child!');
+        child.kill('SIGKILL');
+      }
+
+      proc.execSync('cd ~/launch16/instance/dist && npm install');
+      child = proc.spawn('node', [path.resolve(__dirname, 'dist', 'index.js')]);
+      child.stdout.on('data', data => process.stdout.write(data));
+      child.stderr.on('data', data => process.stderr.write(data));
     }
 
-    console.log('about to spawn node to monitor:', appName);
-    proc.execSync(`cd ~/launch16/instance/dist/${appName} && npm install`);
-    child = proc.spawn('node', [path.resolve(__dirname, `dist/${appName}`, 'index.js')]);
-    pid = child.pid;
-
-    console.log('pid:', pid);
-
-    console.log('isDeploying:', false);
     appRef.child('isDeploying').set(false);
-    child.stdout.on('data', data => process.stdout.write(data));
-    child.stderr.on('data', data => process.stderr.write(data));
   });
 });
